fixes app linking - issue 341
